+++ src/components/Polling/index.tsx
